+++ ledger/participant-state-index/BUILD.bazel
@@ -27,6 +27,7 @@ da_scala_library(
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
         "@maven//:com_typesafe_akka_akka_stream_2_12",
+        "@maven//:io_dropwizard_metrics_metrics_core",
         "@maven//:org_scalaz_scalaz_core_2_12",
     ],
 )
