+++ ledger/participant-state/BUILD.bazel
@@ -26,9 +26,11 @@ da_scala_library(
         "//language-support/scala/bindings",
         "//ledger/ledger-api-health",
         "//ledger/participant-state/protobuf:ledger_configuration_java_proto",
+        "//libs-scala/direct-execution-context",
         "@maven//:com_google_protobuf_protobuf_java",
         "@maven//:com_typesafe_akka_akka_actor_2_12",
         "@maven//:com_typesafe_akka_akka_stream_2_12",
+        "@maven//:io_dropwizard_metrics_metrics_core",
     ],
 )
