+++ ledger/participant-state/src/main/scala/com/daml/ledger/participant/state/v1/TimedReadService.scala
@@ -1,14 +1,12 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.state
+package com.daml.ledger.participant.state.v1
 
 import akka.NotUsed
 import akka.stream.scaladsl.Source
 import com.codahale.metrics.MetricRegistry
-import com.daml.ledger.participant.state.v1.{LedgerInitialConditions, Offset, ReadService, Update}
 import com.digitalasset.ledger.api.health.HealthStatus
-import com.digitalasset.platform.metrics.timedSource
 
 final class TimedReadService(delegate: ReadService, metrics: MetricRegistry, prefix: String)
     extends ReadService {
@@ -22,5 +20,5 @@ final class TimedReadService(delegate: ReadService, metrics: MetricRegistry, pre
     delegate.currentHealth()
 
   private def time[Out, Mat](name: String, source: => Source[Out, Mat]): Source[Out, Mat] =
-    timedSource(metrics.timer(s"$prefix.$name"), source)
+    Metrics.timedSource(metrics.timer(s"$prefix.$name"), source)
 }
