+++ ledger/participant-state/src/main/scala/com/daml/ledger/participant/state/v1/TimedWriteService.scala
@@ -1,25 +1,14 @@
 // Copyright (c) 2020 The DAML Authors. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.digitalasset.platform.state
+package com.daml.ledger.participant.state.v1
 
 import java.util.concurrent.CompletionStage
 
 import com.codahale.metrics.MetricRegistry
-import com.daml.ledger.participant.state.v1.{
-  Configuration,
-  Party,
-  SubmissionId,
-  SubmissionResult,
-  SubmittedTransaction,
-  SubmitterInfo,
-  TransactionMeta,
-  WriteService
-}
 import com.digitalasset.daml.lf.data.Time
 import com.digitalasset.daml_lf_dev.DamlLf
 import com.digitalasset.ledger.api.health.HealthStatus
-import com.digitalasset.platform.metrics.timedFuture
 
 final class TimedWriteService(delegate: WriteService, metrics: MetricRegistry, prefix: String)
     extends WriteService {
@@ -57,5 +46,5 @@ final class TimedWriteService(delegate: WriteService, metrics: MetricRegistry, p
     delegate.currentHealth()
 
   private def time[T](name: String, future: => CompletionStage[T]): CompletionStage[T] =
-    timedFuture(metrics.timer(s"$prefix.$name"), future)
+    Metrics.timedCompletionStage(metrics.timer(s"$prefix.$name"), future)
 }
