+++ ledger/sandbox/src/main/scala/com/digitalasset/platform/metrics/package.scala
@@ -3,23 +3,12 @@
 
 package com.digitalasset.platform
 
-import java.util.concurrent.CompletionStage
-
-import akka.Done
-import akka.stream.scaladsl.{Keep, Source}
 import com.codahale.metrics.Timer
 import com.digitalasset.dec.DirectExecutionContext
 
 import scala.concurrent.Future
 
 package object metrics {
-  def timedFuture[T](timer: Timer, future: => CompletionStage[T]): CompletionStage[T] = {
-    val ctx = timer.time()
-    future.whenComplete { (_, _) =>
-      ctx.stop()
-      ()
-    }
-  }
 
   def timedFuture[T](timer: Timer, future: => Future[T]): Future[T] = {
     val ctx = timer.time()
@@ -28,14 +17,4 @@ package object metrics {
     result
   }
 
-  def timedSource[Out, Mat](timer: Timer, source: => Source[Out, Mat]): Source[Out, Mat] = {
-    val ctx = timer.time()
-    source
-      .watchTermination()(Keep.both[Mat, Future[Done]])
-      .mapMaterializedValue {
-        case (mat, done) =>
-          done.onComplete(_ => ctx.stop())(DirectExecutionContext)
-          mat
-      }
-  }
 }
